node → convert Node to an interface, rename Node to NodeImpl
The public interface Document.getNodes() returned the package-private implementation class Node. This is now an interface with public methods to read the state of the DOM node.
