+++ src/test/java/eu/cessda/cmv/core/ElementInSetValidatorTest.java
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -19,18 +19,18 @@
  */
 package eu.cessda.cmv.core;
 
-import static org.gesis.commons.test.hamcrest.OptionalMatchers.isEmpty;
-import static org.gesis.commons.test.hamcrest.OptionalMatchers.isPresent;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.hasSize;
-import static org.mockito.Mockito.when;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
 
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Set;
 
-import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
+import static org.gesis.commons.test.hamcrest.OptionalMatchers.isEmpty;
+import static org.gesis.commons.test.hamcrest.OptionalMatchers.isPresent;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.hasSize;
+import static org.mockito.Mockito.when;
 
 class ElementInSetValidatorTest
 {
@@ -49,7 +49,7 @@ class ElementInSetValidatorTest
 
 	private Node mockNode( String element )
 	{
-		Node node = Mockito.mock( Node.class );
+		Node node = Mockito.mock( NodeImpl.class );
 		when( node.getTextContent() ).thenReturn( element );
 		return node;
 	}
